Refactor employment-related DTOs and services: rename packages for better organization, update imports, and remove unused entities to enhance code maintainability.
