--- src/main/java/org/example/cyberwatch/features/form/model/AttachmentDTO.java
@@ -1,43 +1,0 @@
-package org.example.cyberwatch.features.form.model;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class AttachmentDTO {
-
-    private Long id;
-
-    @NotBlank(message = "File name cannot be blank")
-    @Size(min = 1, max = 255, message = "File name must be between 1 and 255 characters")
-    private String fileName;
-
-    @NotBlank(message = "Content type cannot be blank")
-    @Size(max = 100, message = "Content type must be max 100 characters")
-    private String contentType;
-
-    @NotNull(message = "File size cannot be null")
-    @Positive(message = "File size must be positive")
-    private Long fileSize;
-
-    @NotBlank(message = "S3 key cannot be blank")
-    @Size(min = 1, max = 500, message = "S3 key must be between 1 and 500 characters")
-    private String s3Key;
-
-    private LocalDateTime uploadDate;
-
-    private Long reportFormId;
-
-}
-
