+++ src/main/java/org/example/cyberwatch/features/form/service/EmploymentFormService.java
@@ -2,8 +2,12 @@ package org.example.cyberwatch.features.form.service;
 
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.RandomStringUtils;
+import org.example.cyberwatch.features.form.dto.CreateEmploymentDTO;
+import org.example.cyberwatch.features.form.dto.EmploymentFormDTO;
+import org.example.cyberwatch.features.form.dto.UpdateEmploymentDTO;
 import org.example.cyberwatch.features.form.exception.EmploymentFormNotFound;
-import org.example.cyberwatch.features.form.model.*;
+import org.example.cyberwatch.features.form.mapper.EmploymentMapper;
+import org.example.cyberwatch.features.form.model.EmploymentForm;
 import org.example.cyberwatch.features.form.repository.EmploymentFormRepository;
 import org.example.cyberwatch.features.staff.exception.StaffNotFoundException;
 import org.example.cyberwatch.features.staff.model.Staff;
