+++ src/main/resources/db/migration/V9__remove_unused_entitys.sql
@@ -1,0 +1,2 @@
+DROP TABLE if EXISTS attachment CASCADE;
+DROP TABLE if EXISTS report_form CASCADE;
